backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hub-manager-backend"
version = "0.1.0"
description = "Hub Manager Backend API"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.18.4",
    "anyio==4.12.1",
    "asyncpg==0.31.0",
    "bcrypt==5.0.0",
    "cryptography==46.0.5",
    "fastapi==0.135.1",
    "httpx==0.28.1",
    "openpyxl==3.1.5",
    "passlib==1.7.4",
    "pillow==12.1.1",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "python-dotenv==1.2.2",
    "python-jose==3.5.0",
    "reportlab==4.4.10",
    "SQLAlchemy==2.0.48",
    "uvicorn==0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "mypy==1.14.1",
    "ruff==0.11.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = [
    ".venv",
    "venv",
    "alembic",
    "__pycache__",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = false
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
exclude = ["alembic/", "venv/", ".venv/"]
